Windows build script (#154)
* Cast from float explicitly.

* Upgrades to waf 2.0.2.

For VS2017 support on Windows.

* [waf] Supports building on Windows.

Needs 7z.exe in PATH for extracting prebuilt clang NSIS installer (when
using bundled clang).
RPATH is "emulated" by putting a libclang.dll alongside the cquery
binary.
Tested with Visual Studio 2017 toolchain.

* [waf] Fix default_resource_directory on Windows + system clang.

* Add AppVeyor configuration

+++ src/messages/initialize.cc
@@ -153,9 +153,9 @@ struct InitializeHandler : BaseMessageHandler<Ipc_InitializeRequest> {
       if (config->indexerCount == 0) {
         // If the user has not specified how many indexers to run, try to
         // guess an appropriate value. Default to 80% utilization.
-        const float kDefaultTargetUtilization = 0.8;
-        config->indexerCount =
-            std::thread::hardware_concurrency() * kDefaultTargetUtilization;
+        const float kDefaultTargetUtilization = 0.8f;
+        config->indexerCount = (int)(
+            std::thread::hardware_concurrency() * kDefaultTargetUtilization);
         if (config->indexerCount <= 0)
           config->indexerCount = 1;
       }
